refactor config.py to separate target specific code
- split config.py to DeployConfig base class and Ec2DeployConfig,
  ExistingDeployConfig, and AzureDeployConfig subclasses
- add @abstractmethod annotations for common interfaces to DeployConfig
  base class
- move target specific code to respective subclasses
